Move performExam TRPC into its own function before refactoring.
